+++ apis/r/NEWS.md
@@ -12,6 +12,7 @@
 * Add new Arrow-to-R type mapper
 * Add transitional/non-exported `parse_query_condition_new` [#3162](https://github.com/single-cell-data/TileDB-SOMA/pull/3162)
 * Apply new `parse_query_condition` [#3174](https://github.com/single-cell-data/TileDB-SOMA/pull/3174)
+* Apply new `non_empty_domain` [#3176](https://github.com/single-cell-data/TileDB-SOMA/pull/3176)
 
 # tiledbsoma 1.14.1